+++ src/utils/generateUrlScheme.js
@@ -3,7 +3,7 @@
  * @param {string} url
  * @param {'youtube'} scheme
  */
-export default function generateUrlScheme(url, scheme) {
+export function generateUrlScheme(url, scheme) {
   const updatedUrl = url.replace(/^(https?:\/\/)/, "")
   return `${scheme}://${updatedUrl}`
 }
