+++ src/utils/isRedditUrl.js
@@ -2,7 +2,7 @@
  *
  * @param {string} url
  */
-export default function isRedditUrl(url) {
+export function isRedditUrl(url) {
   const regex = /^(https?:\/\/)?(www\.)?(reddit\.com|old\.reddit\.com)\/.+/
   return regex.test(url)
 }
