+++ src/utils/isYoutubeUrl.js
@@ -2,7 +2,7 @@
  *
  * @param {string} url
  */
-export default function isYoutubeUrl(url) {
+export function isYoutubeUrl(url) {
   const regex = /^(https?:\/\/)?(www\.)?(youtube\.com|youtu\.be)\/.+/
   return regex.test(url)
 }
